Fix bugs in unpack200
* Do correct endian detection on BSD when defining swap macros.  BSD
  defines both _LITTLE_ENDIAN and _BIG_ENDIAN on all arches, so it is
  insufficient to check for the definition of it there.

Without this fix unpack200 always produces the error:

CRC error, invalid compressed data.

Note that this is not a direct forward port from openjdk8.  That uses a
separate VM_{BIG,LITTLE}_ENDIAN define, but the openjdk11 source is less
consistent with both defining this during the build and in applying it to
the source.  It's probably time with the more modular build to rethink
the trade offs between the previous approach or doing a lengthier code
changes like this.