Sync with JDK 11
This prevents a compilation failure on ARM based architectures.
A better fix would be to remove this and stop redefining

PNG_ARM_NEON_OPT

by not explicitly defining it in CFLAGS.  This should be done
and propagated through versions appropriately and this change
removed.  Overall that needs no definition when using an external
libpng and libpng16.